+++ exporters/prometheus/test/prometheus_test_helper.h
@@ -23,6 +23,10 @@ struct TestDataPoints
   Resource resource = Resource::Create(ResourceAttributes{});
   nostd::unique_ptr<InstrumentationScope> instrumentation_scope =
       InstrumentationScope::Create("library_name", "1.2.0");
+  opentelemetry::common::SystemTimestamp start_ts = 
+    opentelemetry::common::SystemTimestamp{std::chrono::microseconds{1766662500000}};
+  opentelemetry::common::SystemTimestamp end_ts = 
+    opentelemetry::common::SystemTimestamp{std::chrono::microseconds{1766662560000}};
 
   /**
    * Helper function to create ResourceMetrics
@@ -39,8 +43,9 @@ struct TestDataPoints
         metric_sdk::InstrumentDescriptor{"library_name", "description", "unit",
                                          metric_sdk::InstrumentType::kCounter,
                                          metric_sdk::InstrumentValueType::kDouble},
-        metric_sdk::AggregationTemporality::kDelta, opentelemetry::common::SystemTimestamp{},
-        opentelemetry::common::SystemTimestamp{},
+        metric_sdk::AggregationTemporality::kDelta,
+        start_ts,
+        end_ts,
         std::vector<metric_sdk::PointDataAttributes>{
             {metric_sdk::PointAttributes{{"a1", "b1"}}, sum_point_data},
             {metric_sdk::PointAttributes{{"a2", "b2"}}, sum_point_data2}}};
@@ -67,13 +72,15 @@ struct TestDataPoints
         metric_sdk::InstrumentDescriptor{"library_name", "description", "unit",
                                          metric_sdk::InstrumentType::kHistogram,
                                          metric_sdk::InstrumentValueType::kDouble},
-        metric_sdk::AggregationTemporality::kDelta, opentelemetry::common::SystemTimestamp{},
-        opentelemetry::common::SystemTimestamp{},
+        metric_sdk::AggregationTemporality::kDelta,
+        start_ts,
+        end_ts,
         std::vector<metric_sdk::PointDataAttributes>{
             {metric_sdk::PointAttributes{{"a1", "b1"}}, histogram_point_data},
             {metric_sdk::PointAttributes{{"a2", "b2"}}, histogram_point_data2}}};
     data.scope_metric_data_ = std::vector<metric_sdk::ScopeMetrics>{
         {instrumentation_scope.get(), std::vector<metric_sdk::MetricData>{metric_data}}};
+
     return data;
   }
 
@@ -93,8 +100,9 @@ struct TestDataPoints
         metric_sdk::InstrumentDescriptor{"library_name", "description", "unit",
                                          metric_sdk::InstrumentType::kCounter,
                                          metric_sdk::InstrumentValueType::kDouble},
-        metric_sdk::AggregationTemporality::kDelta, opentelemetry::common::SystemTimestamp{},
-        opentelemetry::common::SystemTimestamp{},
+        metric_sdk::AggregationTemporality::kDelta,
+        start_ts,
+        end_ts,
         std::vector<metric_sdk::PointDataAttributes>{
             {metric_sdk::PointAttributes{{"a1", "b1"}}, last_value_point_data},
             {metric_sdk::PointAttributes{{"a2", "b2"}}, last_value_point_data2}}};
@@ -113,8 +121,9 @@ struct TestDataPoints
         metric_sdk::InstrumentDescriptor{"library_name", "description", "unit",
                                          metric_sdk::InstrumentType::kCounter,
                                          metric_sdk::InstrumentValueType::kDouble},
-        metric_sdk::AggregationTemporality::kDelta, opentelemetry::common::SystemTimestamp{},
-        opentelemetry::common::SystemTimestamp{},
+        metric_sdk::AggregationTemporality::kDelta,
+        start_ts,
+        end_ts,
         std::vector<metric_sdk::PointDataAttributes>{
             {metric_sdk::PointAttributes{{"a1", "b1"}}, drop_point_data},
             {metric_sdk::PointAttributes{{"a2", "b2"}}, drop_point_data2}}};